Refactor serial class into separate modules
