+++ bk7231tools/serial/__init__.py
@@ -1,40 +1,14 @@
 from binascii import crc32
-from io import BytesIO
-from math import ceil
-from time import sleep
-from typing import IO, Generator
+from typing import IO
 
-from serial import Serial, Timeout
+from serial import Serial
 
-from .packets import (
-    BkBootVersionCmnd,
-    BkBootVersionResp,
-    BkCheckCrcCmnd,
-    BkCheckCrcResp,
-    BkFlashEraseBlockCmnd,
-    BkFlashRead4KCmnd,
-    BkFlashRead4KResp,
-    BkFlashWrite4KCmnd,
-    BkFlashWriteCmnd,
-    BkFlashWriteResp,
-    BkLinkCheckCmnd,
-    BkLinkCheckResp,
-    BkReadRegCmnd,
-    BkReadRegResp,
-    BkRebootCmnd,
-    BkSetBaudRateCmnd,
-    EraseSize,
-)
-from .protocol import BK7231Protocol
+from .cmd_flash import BK7231CmdFlash
+from .packets import EraseSize
+from .utils import fix_addr
 
 
-def fix_addr(addr: int) -> int:
-    addr &= 0x1FFFFF
-    addr |= 0x200000
-    return addr
-
-
-class BK7231Serial(BK7231Protocol):
+class BK7231Serial(BK7231CmdFlash):
     chip_info: str
     crc_end_incl: bool = False
     crc_speed_bps: int = 400000
@@ -80,185 +54,6 @@ class BK7231Serial(BK7231Protocol):
         if self.serial and not self.serial.closed:
             self.serial.close()
             self.serial = None
-
-    def wait_for_link(self, timeout: float) -> bool:
-        tm = Timeout(timeout)
-        tm_prev = self.serial.timeout
-        self.serial.timeout = 0.005
-
-        command = BkLinkCheckCmnd()
-        response = None
-        while not tm.expired() and not response:
-            try:
-                response: BkLinkCheckResp = self.command(command)
-                if response and response.value != 0:
-                    response = None
-            except ValueError:
-                pass
-
-        self.drain()
-        self.serial.timeout = tm_prev
-        return not not response
-
-    def set_baudrate(self, baudrate: int) -> bool:
-        command = BkSetBaudRateCmnd(baudrate, delay_ms=20)
-
-        def baudrate_cb():
-            if self.debug_hl:
-                print("-- UART: Changing port baudrate")
-            sleep(command.delay_ms / 1000 / 2)
-            self.serial.baudrate = baudrate
-
-        self.command(command, after_send=baudrate_cb)
-        return True
-
-    def reboot_chip(self):
-        command = BkRebootCmnd()
-        self.command(command)
-
-    def read_chip_info(self) -> str:
-        command = BkBootVersionCmnd()
-        response: BkBootVersionResp = self.command(command)
-        if response.version == b"\x07":
-            # read chip type from register if command is not implemented
-            self.chip_info = hex(self.register_read(0x800000))  # SCTRL_CHIP_ID
-        else:
-            self.chip_info = response.version
-        return self.chip_info
-
-    def register_read(self, address: int) -> int:
-        command = BkReadRegCmnd(address)
-        response: BkReadRegResp = self.command(command)
-        return response.value
-
-    def read_flash_range_crc(self, start: int, end: int) -> int:
-        start = fix_addr(start)
-        end = fix_addr(end)
-        # probably reading whole flash CRC
-        if end == 0x200000:
-            end += 0x200000
-        # command arguments are (incl., excl.)
-        if start == end:
-            raise ValueError("Start and end must differ! (end is exclusive)")
-        # print a warning instead of just timeout-ing
-        timeout_current = self.serial.timeout
-        timeout_minimum = (end - start) / self.crc_speed_bps
-        if timeout_minimum > timeout_current:
-            print(
-                "WARN: The current command timeout of",
-                timeout_current,
-                "second(s) is too low for reading",
-                end - start,
-                "bytes CRC. Increasing to",
-                ceil(timeout_minimum),
-                "second(s).",
-            )
-            self.serial.timeout = ceil(timeout_minimum)
-        # fix for BK7231N which also counts the end offset
-        if self.crc_end_incl:
-            end -= 1
-        command = BkCheckCrcCmnd(start, end)
-        response: BkCheckCrcResp = self.command(command)
-        self.serial.timeout = timeout_current
-        return response.crc32 ^ 0xFFFFFFFF
-
-    def check_crc(self, start: int, data: bytes):
-        chip = self.read_flash_range_crc(start, start + len(data))
-        calc = crc32(data)
-        if chip != calc:
-            raise ValueError(
-                f"Chip CRC value {chip:X} does not match calculated CRC value {calc:X}"
-            )
-
-    def read_flash_4k(
-        self,
-        start: int,
-        count: int = 1,
-        crc_check: bool = True,
-    ) -> bytes:
-        out = BytesIO()
-        for data in self.read_flash(start, count * 4096, crc_check):
-            out.write(data)
-        return out.getvalue()
-
-    def read_flash(
-        self,
-        start: int,
-        length: int,
-        crc_check: bool = True,
-    ) -> Generator[bytes, None, None]:
-        start = fix_addr(start)
-        if start & 0xFFF:
-            raise ValueError(f"Starting address 0x{start:X} is not 4K aligned")
-        if length & 0xFFF:
-            raise ValueError(f"Read length 0x{length:X} is not 4K aligned")
-        length = int(length // 4096)
-
-        for i in range(length):
-            addr = start + i * 4096
-            progress = i / length * 100.0
-            print(f"Reading 4k page at 0x{addr:06X} ({progress:.2f}%)")
-            command = BkFlashRead4KCmnd(addr)
-            response: BkFlashRead4KResp = self.command(command)
-            if crc_check:
-                self.check_crc(addr, response.data)
-            yield response.data
-
-    def erase_flash_block(
-        self,
-        start: int,
-        size: EraseSize,
-        dry_run: bool = False,
-    ) -> bool:
-        start = fix_addr(start)
-        if dry_run:
-            print(f" -> would erase {size.name} at 0x{start:X}")
-            return True
-        command = BkFlashEraseBlockCmnd(size, start)
-        self.command(command)
-        return True
-
-    def write_flash_bytes(
-        self,
-        start: int,
-        data: bytes,
-        crc_check: bool = False,
-        dry_run: bool = False,
-    ) -> bool:
-        start = fix_addr(start)
-        if len(data) > 256:
-            raise ValueError(f"Data too long ({len(data)} > 256)")
-        if dry_run:
-            print(f" -> would write {len(data)} bytes to 0x{start:X}")
-            return True
-        command = BkFlashWriteCmnd(start, data)
-        response: BkFlashWriteResp = self.command(command)
-        if response.written != len(data):
-            raise ValueError(f"Writing failed; wrote only {response.written} bytes")
-        if crc_check:
-            self.check_crc(start, data)
-        return True
-
-    def write_flash_4k(
-        self,
-        start: int,
-        data: bytes,
-        crc_check: bool = False,
-        dry_run: bool = False,
-    ) -> bool:
-        start = fix_addr(start)
-        if len(data) > 4096:
-            raise ValueError(f"Data too long ({len(data)} > 4096)")
-        if len(data) < 4096:
-            data += (4096 - len(data)) * b"\xff"
-        if dry_run:
-            print(f" -> would write {len(data)} bytes to 0x{start:X}")
-            return True
-        command = BkFlashWrite4KCmnd(start, data)
-        self.command(command)
-        if crc_check:
-            self.check_crc(start, data)
-        return True
 
     def program_flash(
         self,
